Jenkinsfile: Fix vivado build info
Fix the vivado build info record in the commitIDs file. Also,
resolve the daily_latest symlink once at the start of the job and
then use the build version when sourcing the tools. This way we
make sure that all stages of the build use the same tools version
as the daily_latest symlink could move while our build is running
and the tools get sourced for each stage as those spawn a new shell.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>